backend_api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

azure-storage-blob==12.19.0

PyPDF2==3.0.1
python-docx==1.1.0

python-dotenv==1.0.0
azure-identity